requirements.txt
Flask==2.2.5
python-dotenv==0.21.0
openai==0.27.0
gunicorn==20.1.0
